--- src/main/java/com/naver/webtoon/common/exception/WebtoonException.java
@@ -1,11 +1,0 @@
-package com.naver.webtoon.common.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class WebtoonException extends RuntimeException{
-
-    private final ErrorCode errorCode;
-}
