--- src/main/java/com/naver/webtoon/common/security/UserDetailsImpl.java
@@ -1,56 +1,0 @@
-package com.naver.webtoon.common.security;
-
-import com.naver.webtoon.member.entity.Member;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.Collection;
-
-public class UserDetailsImpl implements UserDetails {
-
-    private final Member member;
-    private final String username;
-
-    public UserDetailsImpl(Member member, String username) {
-        this.member = member;
-        this.username = username;
-    }
-
-    public Member getMember() {
-        return member;
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return null;
-    }
-
-    @Override
-    public String getUsername() {
-        return this.username;
-    }
-    @Override
-    public String getPassword() {
-        return null;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return false;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return false;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return false;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return false;
-    }
-}
