+++ src/main/java/com/naver/webtoon/domain/member/repository/MemberRepository.java
@@ -1,0 +1,12 @@
+package com.naver.webtoon.domain.member.repository;
+
+import com.naver.webtoon.domain.member.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+    Optional<Member> findByUsername(String username);
+    boolean existsByUsername(String username);
+}
