+++ src/main/java/com/naver/webtoon/global/common/exception/WebtoonException.java
@@ -1,0 +1,11 @@
+package com.naver.webtoon.global.common.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class WebtoonException extends RuntimeException{
+
+    private final ErrorCode errorCode;
+}
