--- src/main/java/com/naver/webtoon/member/repository/MemberRepository.java
@@ -1,12 +1,0 @@
-package com.naver.webtoon.member.repository;
-
-import com.naver.webtoon.member.entity.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-    Optional<Member> findByUsername(String username);
-    boolean existsByUsername(String username);
-}
